Use base class for project list components
