+++ CollAction/Frontend/app/project/index.tsx
@@ -1,5 +1,4 @@
 import * as React from "react";
-import FindProject from "./FindProject";
 
 import "./styles/FindProject.scss";
 import "./styles/ProjectDetails.scss";
@@ -10,20 +9,10 @@ import "./styles/ProjectCommit.scss";
 import "./styles/ProjectThankYouCommit.scss";
 import "./styles/ProjectSendEmail.scss";
 
+import "./FindProject";
+
 import "./PieChart";
 import "./Carousel";
 
 import "./UploadBanner";
 import "./UploadDescriptiveImage";
-
-import renderComponentIf from "../global/renderComponentIf";
-
-renderComponentIf(
-  <FindProject controller={true} />,
-  document.getElementById("project-controller")
-);
-
-renderComponentIf(
-  <FindProject controller={false} />,
-  document.getElementById("projects-container")
-);
